fix: increase polling rate

+++ src/ytb_dl_service/consumer/client.py
@@ -12,32 +12,38 @@ handlers = EventsHandlers()
 
 
 def longer_than_10_minutes(info, *, incomplete):
-    duration = info.get('duration')
+    duration = info.get("duration")
     if duration and duration > 600:
-        return 'The video is too long'
+        return "The video is too long"
 
 
 video_opts = {
     "match_filter": longer_than_10_minutes,
     "format": "best[height<=480]",
-    "outtmpl": "%(title)s.%(ext)s"
+    "outtmpl": "%(title)s.%(ext)s",
 }
 
 audio_opts = {
     "match_filter": longer_than_10_minutes,
-    'format': 'bestaudio/best',
-    'postprocessors': [{
-        'key': 'FFmpegExtractAudio',
-        'preferredcodec': 'opus',
-    }],
-    'outtmpl': "%(title)s.%(ext)s",
+    "format": "bestaudio/best",
+    "postprocessors": [
+        {
+            "key": "FFmpegExtractAudio",
+            "preferredcodec": "opus",
+        }
+    ],
+    "outtmpl": "%(title)s.%(ext)s",
 }
 
 
 async def reply(opts: dict[str, Any], url: str, event: Event):
     with YoutubeDL(opts) as ytdl:
         info = ytdl.extract_info(url, download=True)
-        filepath = filepath = info['requested_downloads'][0]['filepath']
+        if info is None:
+            return await event.reply(
+                json.dumps({"msg": "Error fetching video"}, metadata={"error": "1"})
+            )
+        filepath = info["requested_downloads"][0]["filepath"]
         with open(filepath, "rb") as f:
             await event.reply(f, metadata={"filename": os.path.basename(filepath)})
         Path(filepath).unlink(True)
@@ -49,7 +55,9 @@ async def new_url(event: Event):
     async with aiohttp.ClientSession() as s:
         event.__session = s
         if event.payload is None:
-            return await event.reply(json.dumps({"msg": "Invalid URL"}), metadata={"error": "1"})
+            return await event.reply(
+                json.dumps({"msg": "Invalid URL"}), metadata={"error": "1"}
+            )
         payload: dict[str, str] = json.loads(event.payload)
         url = payload["url"]
         kind: Literal["audio", "video"] = payload["kind"]
@@ -61,12 +69,14 @@ async def new_url(event: Event):
             print(str(e.__class__))
             print(str(e.__cause__))
             print(str(e))
-            return await event.reply(json.dumps({"msg": "Could not download video"}), metadata={"error": "1"})
+            return await event.reply(
+                json.dumps({"msg": "Could not download video"}), metadata={"error": "1"}
+            )
 
 
 async def main():
     async with EventPolling(handlers) as poller:
-        await poller.start(0.1)
+        await poller.start(1)
 
 
 if __name__ == "__main__":
